Split news headline to a now component
